+++ datafusion/ffi/src/execution_plan.rs
@@ -68,7 +68,7 @@ pub struct FFI_ExecutionPlan {
 
     /// Utility to identify when FFI objects are accessed locally through
     /// the foreign interface.
-    pub library_marker_id: extern "C" fn() -> u64,
+    pub library_marker_id: extern "C" fn() -> usize,
 }
 
 unsafe impl Send for FFI_ExecutionPlan {}
